Updates to change 'converter' to 'decorator' and clarify reason for TracerObserverResolver
