--- opentracing-api-extensions-tracerconverter/src/main/resources/META-INF/services/io.opentracing.contrib.tracerresolver.TracerConverter
@@ -1,1 +1,0 @@
-io.opentracing.contrib.api.tracer.converter.APIExtensionsTracerConverter
