--- opentracing-api-extensions-tracerconverter/src/test/java/io/opentracing/contrib/api/tracer/converter/TestTracerObserverResolver.java
@@ -1,31 +1,0 @@
-/**
- * Copyright 2017 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package io.opentracing.contrib.api.tracer.converter;
-
-import io.opentracing.contrib.api.TracerObserver;
-
-public class TestTracerObserverResolver implements TracerObserverResolver {
-
-    private static TracerObserver tracerObserver;
-
-    public static void setTracerObserver(TracerObserver t) {
-        tracerObserver = t;
-    }
-
-    @Override
-    public TracerObserver resolve() {
-        return tracerObserver;
-    }
-
-}
