--- opentracing-api-extensions-tracerconverter/src/test/resources/META-INF/services/io.opentracing.contrib.api.tracer.converter.TracerObserverResolver
@@ -1,1 +1,0 @@
-io.opentracing.contrib.api.tracer.converter.TestTracerObserverResolver
