--- opentracing-api-extensions-tracerconverter/src/test/resources/META-INF/services/io.opentracing.contrib.tracerresolver.TracerResolver
@@ -1,1 +1,0 @@
-io.opentracing.contrib.api.tracer.converter.TestTracerResolver
