+++ opentracing-api-extensions-tracerdecorator/src/main/resources/META-INF/services/io.opentracing.contrib.tracerresolver.TracerConverter
@@ -1,0 +1,1 @@
+io.opentracing.contrib.api.tracer.decorator.APIExtensionsTracerDecorator
