+++ opentracing-api-extensions-tracerdecorator/src/test/resources/META-INF/services/io.opentracing.contrib.api.tracer.decorator.TracerObserverResolver
@@ -1,0 +1,1 @@
+io.opentracing.contrib.api.tracer.decorator.TestTracerObserverResolver
