+++ opentracing-api-extensions-tracerdecorator/src/test/resources/META-INF/services/io.opentracing.contrib.tracerresolver.TracerResolver
@@ -1,0 +1,1 @@
+io.opentracing.contrib.api.tracer.decorator.TestTracerResolver
